backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.1.0

# HTTP and API
httpx>=0.26.0
aiofiles>=23.2.1

# AI Integration
anthropic>=0.18.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
